Implement list events, create events, purchase, trasnfer and check in

+++ assembly/models.ts
@@ -2,11 +2,11 @@ import {PersistentMap, PersistentVector, u128} from "near-sdk-as";
 
 @nearBindgen
 export class Ticket {
-    purchased_at: u128
+    purchased_at: u64
     owner: string
     price: u128
     check_in: boolean
-    check_in_at: u128
+    check_in_at: u64
 }
 
 @nearBindgen
@@ -17,10 +17,9 @@ export class Event {
     symbol: string;
     occupied: i32;
     seatPrice: u128;
-    start: u128;
-    end: u128;
+    start: u64;
+    end: u64;
     initialSupply: i32;
     attendees: PersistentVector<string>
     tickets: PersistentMap<string, Ticket>
-
 }
